+++ Assets/Scripts/Interfaces/IStorage.cs
@@ -2,7 +2,7 @@
 {
     public interface IStorage<T>
     {
-        public T Get(int n);
-        public void Add(int n);
+        public T Get(float n);
+        public void Add(float n);
     }
 }
